Adding new features such as utils, configs, database and model (#1)
* add cache, config and utils

* add model and database

* modified app, config,database and model

* :fire: refactor: update import statement to use model for response generation

* :memo: :fire: feat: implement caching for database schema and model loading

* :zap::fire: refactor: remove dotenv loading and unused model path configuration

* :fire: refactor: remove unused database schema caching and related functions

* :zap::fire: refactor: enhance response generation by optimizing database queries and caching

* :fire::zap: feat: load environment variables from .env file for configuration

---------
